Mobile App
The website also has a search bar, which allows users to search for specific products. There is also a shopping cart icon, which allows users to view their cart and proceed to checkout.

Homepage: The homepage of the website is designed to be informative and engaging. It features a banner image with a promotional offer, as well as a variety of products from different categories. This allows users to quickly get an overview of the website and the products that it offers.

Product categories: The website has three main product categories Men, Women, and Kids. Each category has a sub-category, which allows users to narrow down their search for specific products. For example, the Men category has sub-categories for Shirts, Pants, and Shoes.

Product pages: Product pages are designed to be informative and visually appealing. They feature a large product image, as well as a detailed description of the product. There is also a section for customer reviews, which allows users to see what other customers have to say about the product.

Shopping cart: The shopping cart icon is located in the top right corner of the website. It allows users to view their cart and proceed to checkout. The shopping cart page shows a list of all of the items in the cart, as well as the total price. Users can then enter their shipping and billing information and proceed to checkout.